Rename `web/pdf_viewer.js` to `web/base_viewer.js`
Please note that the only reason for this change is to try and improve reviewability of later patches, by keeping the diffs more manageable.
